tx: move signing related code into sign mod
